feat: adjust the fallback value for node name
Added the string replace to replace "-" with "_". Whenever the node name is not specified the parser tries to set the node id as the node name. Hence ndoe id is the fallback value. But the node id is usually something like "sid-E75E7BB4-DA1C-4D89-B328-C5F152BF1C47". Such an id is not handled correctly by the solc compiler and results in compilation error. However when the node id is represented as "sid_E75E7BB4_DA1C_4D89_B328_C5F152BF1C47" the solc compiler is able to compile the input json.

Best case scenario is when the user models a diagram he specifies the name for every node and gateways.